requirements.txt
google-api-python-client
youtube-transcript-api
sentence-transformers
scikit-learn
pydantic
pyyaml
requests
python-dotenv
groq
google-generativeai
dnspython
torch
